pyproject.toml
[project]
name = "financialdataextraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.5",
    "google-api-python-client>=2.189.0",
    "google-auth-oauthlib>=1.2.4",
    "openai>=2.17.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pgvector>=0.4.2",
    "plotly>=6.5.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "streamlit>=1.54.0",
    "tabulate>=0.9.0",
    "uvicorn[standard]>=0.40.0",
]
